+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+- Improved documentation
+- Show dashboard widget only to admins
+- Fixed group flushing when group happends to end with blog id
+
 ## 2.2.4
 
 - Register `wp redis` CLI command late