+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Show dashboard widget only to admins
 - Fixed group flushing when group happends to end with blog id
 - Support loading Predis from `WP_REDIS_PLUGIN_PATH`
+- Throw errors when connection error occur
 
 ## 2.2.4